Move cli alongside other Typer modules
