+++ data_safe_haven/commands/__init__.py
@@ -1,13 +1,3 @@
-from .admin import admin_command_group
-from .config import config_command_group
-from .context import context_command_group
-from .deploy import deploy_command_group
-from .teardown import teardown_command_group
+from .cli import application
 
-__all__ = [
-    "admin_command_group",
-    "context_command_group",
-    "config_command_group",
-    "deploy_command_group",
-    "teardown_command_group",
-]
+__all__ = ["application"]
